fix: locale date format bug and timezone bug
`toLocaleDateString` converts dates in different formats depending on the user's locale:

en-US: "12/20/2012"
en-GB: "20/12/2012"

This would fail for users with the en-US locale since we are assuming day-month-year order.

Also fixed a timezone issue, where users were shown expense groups from ±1 day. This was happening since we were sending date filters in the local timezone when the endpoint is expecting it in UTC.